Modified the formatting behavior to be less agressive on input strings that begin with 1. Previously, the formatter would insert the parenthesis around the area code once the string beginning with 1 had a length of 2. Now, the parenthesis are inserted around the area code only once the length of the string beginning with 1 is 5.
This causes better behavior when formatting phone number strings as they are typed. If you set up a UITextField to format the number after each time the string changes. It becomes impossible to delete the string '1 (234)' because the formatter will reinsert the deleted ')'.

Additionally, this commit modifies the tests project to accept this behavior. All other unrelated tests pass.